Add command completion generation to all "user facing" tools
