+++ kanidm_tools/build.rs
@@ -5,6 +5,8 @@ use structopt::clap::Shell;
 use structopt::StructOpt;
 
 include!("src/opt/ssh_authorizedkeys.rs");
+include!("src/opt/badlist_preprocess.rs");
+include!("src/opt/kanidm.rs");
 
 fn main() {
     let outdir = match env::var_os("OUT_DIR") {
@@ -13,15 +15,23 @@ fn main() {
     };
 
     SshAuthorizedOpt::clap().gen_completions(
-        "ssh_authorizedkeys_direct",
+        "kanidm_ssh_authorizedkeys_direct",
         Shell::Bash,
         outdir.clone(),
     );
     SshAuthorizedOpt::clap().gen_completions(
-        "ssh_authorizedkeys_direct",
+        "kanidm_ssh_authorizedkeys_direct",
         Shell::Zsh,
         outdir.clone(),
     );
 
-    println!("{:?}", outdir);
+    BadlistProcOpt::clap().gen_completions(
+        "kanidm_badlist_preprocess",
+        Shell::Bash,
+        outdir.clone(),
+    );
+    BadlistProcOpt::clap().gen_completions("kanidm_badlist_preprocess", Shell::Zsh, outdir.clone());
+
+    KanidmClientOpt::clap().gen_completions("kanidm", Shell::Bash, outdir.clone());
+    KanidmClientOpt::clap().gen_completions("kanidm", Shell::Zsh, outdir.clone());
 }
