+++ kanidm_tools/src/cli/account.rs
@@ -1,181 +1,13 @@
-use crate::common::CommonOpt;
 use crate::password_prompt;
+use crate::{
+    AccountCredential, AccountOpt, AccountPosix, AccountRadius, AccountSsh, AccountValidity,
+};
 use qrcode::render::unicode;
 use qrcode::QrCode;
 use std::io;
-use structopt::StructOpt;
 use time::OffsetDateTime;
 
 use webauthn_authenticator_rs::{u2fhid::U2FHid, WebauthnAuthenticator};
-
-#[derive(Debug, StructOpt)]
-pub struct AccountCommonOpt {
-    #[structopt()]
-    account_id: String,
-}
-
-#[derive(Debug, StructOpt)]
-pub struct AccountCredentialSet {
-    #[structopt(flatten)]
-    aopts: AccountCommonOpt,
-    #[structopt(flatten)]
-    copt: CommonOpt,
-}
-
-#[derive(Debug, StructOpt)]
-pub struct AccountNamedOpt {
-    #[structopt(flatten)]
-    aopts: AccountCommonOpt,
-    #[structopt(flatten)]
-    copt: CommonOpt,
-}
-
-#[derive(Debug, StructOpt)]
-pub struct AccountNamedExpireDateTimeOpt {
-    #[structopt(flatten)]
-    aopts: AccountCommonOpt,
-    #[structopt(flatten)]
-    copt: CommonOpt,
-    #[structopt(name = "datetime")]
-    /// An rfc3339 time of the format "YYYY-MM-DDTHH:MM:SS+TZ", "2020-09-25T11:22:02+10:00"
-    /// or the word "never", "clear" to remove account expiry.
-    datetime: String,
-}
-
-#[derive(Debug, StructOpt)]
-pub struct AccountNamedValidDateTimeOpt {
-    #[structopt(flatten)]
-    aopts: AccountCommonOpt,
-    #[structopt(flatten)]
-    copt: CommonOpt,
-    #[structopt(name = "datetime")]
-    /// An rfc3339 time of the format "YYYY-MM-DDTHH:MM:SS+TZ", "2020-09-25T11:22:02+10:00"
-    /// or the word "any", "clear" to remove valid from enforcement.
-    datetime: String,
-}
-
-#[derive(Debug, StructOpt)]
-pub struct AccountNamedTagOpt {
-    #[structopt(flatten)]
-    aopts: AccountCommonOpt,
-    #[structopt(flatten)]
-    copt: CommonOpt,
-    #[structopt(name = "tag")]
-    tag: String,
-}
-
-#[derive(Debug, StructOpt)]
-pub struct AccountNamedTagPKOpt {
-    #[structopt(flatten)]
-    aopts: AccountCommonOpt,
-    #[structopt(flatten)]
-    copt: CommonOpt,
-    #[structopt(name = "tag")]
-    tag: String,
-    #[structopt(name = "pubkey")]
-    pubkey: String,
-}
-
-#[derive(Debug, StructOpt)]
-pub struct AccountCreateOpt {
-    #[structopt(flatten)]
-    aopts: AccountCommonOpt,
-    #[structopt(name = "display_name")]
-    display_name: String,
-    #[structopt(flatten)]
-    copt: CommonOpt,
-}
-
-#[derive(Debug, StructOpt)]
-pub enum AccountCredential {
-    #[structopt(name = "set_password")]
-    SetPassword(AccountCredentialSet),
-    #[structopt(name = "generate_password")]
-    GeneratePassword(AccountCredentialSet),
-    #[structopt(name = "register_webauthn")]
-    RegisterWebauthn(AccountNamedTagOpt),
-    /// Set the TOTP credential of the account. If a TOTP already exists, on a successful
-    /// registration, this will replace it.
-    #[structopt(name = "set_totp")]
-    RegisterTOTP(AccountNamedTagOpt),
-    /// Remove TOTP from the account. If no TOTP exists, no action is taken.
-    #[structopt(name = "remove_totp")]
-    RemoveTOTP(AccountNamedOpt),
-}
-
-#[derive(Debug, StructOpt)]
-pub enum AccountRadius {
-    #[structopt(name = "show_secret")]
-    Show(AccountNamedOpt),
-    #[structopt(name = "generate_secret")]
-    Generate(AccountNamedOpt),
-    #[structopt(name = "delete_secret")]
-    Delete(AccountNamedOpt),
-}
-
-#[derive(Debug, StructOpt)]
-pub struct AccountPosixOpt {
-    #[structopt(flatten)]
-    aopts: AccountCommonOpt,
-    #[structopt(long = "gidnumber")]
-    gidnumber: Option<u32>,
-    #[structopt(long = "shell")]
-    shell: Option<String>,
-    #[structopt(flatten)]
-    copt: CommonOpt,
-}
-
-#[derive(Debug, StructOpt)]
-pub enum AccountPosix {
-    #[structopt(name = "show")]
-    Show(AccountNamedOpt),
-    #[structopt(name = "set")]
-    Set(AccountPosixOpt),
-    #[structopt(name = "set_password")]
-    SetPassword(AccountNamedOpt),
-}
-
-#[derive(Debug, StructOpt)]
-pub enum AccountSsh {
-    #[structopt(name = "list_publickeys")]
-    List(AccountNamedOpt),
-    #[structopt(name = "add_publickey")]
-    Add(AccountNamedTagPKOpt),
-    #[structopt(name = "delete_publickey")]
-    Delete(AccountNamedTagOpt),
-}
-
-#[derive(Debug, StructOpt)]
-pub enum AccountValidity {
-    #[structopt(name = "show")]
-    Show(AccountNamedOpt),
-    #[structopt(name = "expire_at")]
-    ExpireAt(AccountNamedExpireDateTimeOpt),
-    #[structopt(name = "begin_from")]
-    BeginFrom(AccountNamedValidDateTimeOpt),
-}
-
-#[derive(Debug, StructOpt)]
-pub enum AccountOpt {
-    #[structopt(name = "credential")]
-    Credential(AccountCredential),
-    #[structopt(name = "radius")]
-    Radius(AccountRadius),
-    #[structopt(name = "posix")]
-    Posix(AccountPosix),
-    #[structopt(name = "ssh")]
-    Ssh(AccountSsh),
-    #[structopt(name = "list")]
-    List(CommonOpt),
-    #[structopt(name = "get")]
-    Get(AccountNamedOpt),
-    #[structopt(name = "create")]
-    Create(AccountCreateOpt),
-    #[structopt(name = "delete")]
-    Delete(AccountNamedOpt),
-    #[structopt(name = "validity")]
-    Validity(AccountValidity),
-}
 
 impl AccountOpt {
     pub fn debug(&self) -> bool {
