+++ kanidm_tools/src/cli/group.rs
@@ -1,55 +1,4 @@
-use crate::common::{CommonOpt, Named};
-use structopt::StructOpt;
-
-#[derive(Debug, StructOpt)]
-pub struct GroupNamedMembers {
-    #[structopt()]
-    name: String,
-    #[structopt()]
-    members: Vec<String>,
-    #[structopt(flatten)]
-    copt: CommonOpt,
-}
-
-#[derive(Debug, StructOpt)]
-pub struct GroupPosixOpt {
-    #[structopt()]
-    name: String,
-    #[structopt(long = "gidnumber")]
-    gidnumber: Option<u32>,
-    #[structopt(flatten)]
-    copt: CommonOpt,
-}
-
-#[derive(Debug, StructOpt)]
-pub enum GroupPosix {
-    #[structopt(name = "show")]
-    Show(Named),
-    #[structopt(name = "set")]
-    Set(GroupPosixOpt),
-}
-
-#[derive(Debug, StructOpt)]
-pub enum GroupOpt {
-    #[structopt(name = "list")]
-    List(CommonOpt),
-    #[structopt(name = "get")]
-    Get(Named),
-    #[structopt(name = "create")]
-    Create(Named),
-    #[structopt(name = "delete")]
-    Delete(Named),
-    #[structopt(name = "list_members")]
-    ListMembers(Named),
-    #[structopt(name = "set_members")]
-    SetMembers(GroupNamedMembers),
-    #[structopt(name = "purge_members")]
-    PurgeMembers(Named),
-    #[structopt(name = "add_members")]
-    AddMembers(GroupNamedMembers),
-    #[structopt(name = "posix")]
-    Posix(GroupPosix),
-}
+use crate::{GroupOpt, GroupPosix};
 
 impl GroupOpt {
     pub fn debug(&self) -> bool {
