+++ kanidm_tools/src/cli/lib.rs
@@ -10,7 +10,10 @@
 
 #[macro_use]
 extern crate log;
+use std::path::PathBuf;
 use structopt::StructOpt;
+
+include!("../opt/kanidm.rs");
 
 pub mod account;
 pub mod common;
@@ -18,23 +21,6 @@ pub mod group;
 pub mod login;
 pub mod raw;
 pub mod recycle;
-
-use crate::account::AccountOpt;
-use crate::common::CommonOpt;
-use crate::group::GroupOpt;
-use crate::login::LoginOpt;
-use crate::raw::RawOpt;
-use crate::recycle::RecycleOpt;
-
-#[derive(Debug, StructOpt)]
-pub enum SelfOpt {
-    #[structopt(name = "whoami")]
-    /// Show the current authenticated user's identity
-    Whoami(CommonOpt),
-    #[structopt(name = "set_password")]
-    /// Set the current user's password
-    SetPassword(CommonOpt),
-}
 
 impl SelfOpt {
     pub fn debug(&self) -> bool {
@@ -80,49 +66,26 @@ impl SelfOpt {
     }
 }
 
-#[derive(Debug, StructOpt)]
-#[structopt(about = "Kanidm Client Utility")]
-pub enum ClientOpt {
-    #[structopt(name = "login")]
-    /// Login to an account to use with future cli operations
-    Login(LoginOpt),
-    #[structopt(name = "self")]
-    /// Actions for the current authenticated account
-    CSelf(SelfOpt),
-    #[structopt(name = "account")]
-    /// Account operations
-    Account(AccountOpt),
-    #[structopt(name = "group")]
-    /// Group operations
-    Group(GroupOpt),
-    #[structopt(name = "recycle_bin")]
-    /// Recycle Bin operations
-    Recycle(RecycleOpt),
-    #[structopt(name = "raw")]
-    /// Unsafe - low level, raw database operations.
-    Raw(RawOpt),
-}
-
-impl ClientOpt {
+impl KanidmClientOpt {
     pub fn debug(&self) -> bool {
         match self {
-            ClientOpt::Raw(ropt) => ropt.debug(),
-            ClientOpt::Login(lopt) => lopt.debug(),
-            ClientOpt::CSelf(csopt) => csopt.debug(),
-            ClientOpt::Account(aopt) => aopt.debug(),
-            ClientOpt::Group(gopt) => gopt.debug(),
-            ClientOpt::Recycle(ropt) => ropt.debug(),
+            KanidmClientOpt::Raw(ropt) => ropt.debug(),
+            KanidmClientOpt::Login(lopt) => lopt.debug(),
+            KanidmClientOpt::CSelf(csopt) => csopt.debug(),
+            KanidmClientOpt::Account(aopt) => aopt.debug(),
+            KanidmClientOpt::Group(gopt) => gopt.debug(),
+            KanidmClientOpt::Recycle(ropt) => ropt.debug(),
         }
     }
 
     pub fn exec(&self) {
         match self {
-            ClientOpt::Raw(ropt) => ropt.exec(),
-            ClientOpt::Login(lopt) => lopt.exec(),
-            ClientOpt::CSelf(csopt) => csopt.exec(),
-            ClientOpt::Account(aopt) => aopt.exec(),
-            ClientOpt::Group(gopt) => gopt.exec(),
-            ClientOpt::Recycle(ropt) => ropt.exec(),
+            KanidmClientOpt::Raw(ropt) => ropt.exec(),
+            KanidmClientOpt::Login(lopt) => lopt.exec(),
+            KanidmClientOpt::CSelf(csopt) => csopt.exec(),
+            KanidmClientOpt::Account(aopt) => aopt.exec(),
+            KanidmClientOpt::Group(gopt) => gopt.exec(),
+            KanidmClientOpt::Recycle(ropt) => ropt.exec(),
         }
     }
 }
