+++ kanidm_tools/src/cli/login.rs
@@ -1,4 +1,4 @@
-use crate::common::CommonOpt;
+use crate::LoginOpt;
 use kanidm_client::{ClientError, KanidmClient};
 use kanidm_proto::v1::{AuthAllowed, AuthResponse, AuthState};
 use libc::umask;
@@ -6,7 +6,6 @@ use std::collections::BTreeMap;
 use std::fs::{create_dir, File};
 use std::io::{self, BufReader, BufWriter};
 use std::path::PathBuf;
-use structopt::StructOpt;
 use webauthn_authenticator_rs::{u2fhid::U2FHid, RequestChallengeResponse, WebauthnAuthenticator};
 
 static TOKEN_DIR: &str = "~/.cache";
@@ -100,14 +99,6 @@ fn get_index_choice(len: usize) -> Result<u8, ClientError> {
             Err(_) => eprintln!("Invalid Number"),
         };
     }
-}
-
-#[derive(Debug, StructOpt)]
-pub struct LoginOpt {
-    #[structopt(flatten)]
-    pub copt: CommonOpt,
-    #[structopt(short = "w", long = "webauthn")]
-    pub webauthn: bool,
 }
 
 impl LoginOpt {
