+++ kanidm_tools/src/cli/raw.rs
@@ -1,53 +1,13 @@
-use crate::common::CommonOpt;
+use crate::RawOpt;
 use kanidm_proto::v1::{Entry, Filter, Modify, ModifyList};
 use std::collections::BTreeMap;
-use structopt::StructOpt;
 
 use std::error::Error;
 use std::fs::File;
 use std::io::BufReader;
 use std::path::Path;
-use std::path::PathBuf;
 
 use serde::de::DeserializeOwned;
-
-#[derive(Debug, StructOpt)]
-pub struct FilterOpt {
-    #[structopt()]
-    filter: String,
-    #[structopt(flatten)]
-    commonopts: CommonOpt,
-}
-
-#[derive(Debug, StructOpt)]
-pub struct CreateOpt {
-    #[structopt(parse(from_os_str))]
-    file: PathBuf,
-    #[structopt(flatten)]
-    commonopts: CommonOpt,
-}
-
-#[derive(Debug, StructOpt)]
-pub struct ModifyOpt {
-    #[structopt(flatten)]
-    commonopts: CommonOpt,
-    #[structopt()]
-    filter: String,
-    #[structopt(parse(from_os_str))]
-    file: PathBuf,
-}
-
-#[derive(Debug, StructOpt)]
-pub enum RawOpt {
-    #[structopt(name = "search")]
-    Search(FilterOpt),
-    #[structopt(name = "create")]
-    Create(CreateOpt),
-    #[structopt(name = "modify")]
-    Modify(ModifyOpt),
-    #[structopt(name = "delete")]
-    Delete(FilterOpt),
-}
 
 fn read_file<T: DeserializeOwned, P: AsRef<Path>>(path: P) -> Result<T, Box<dyn Error>> {
     let f = File::open(path)?;
