+++ kanidm_tools/src/cli/recycle.rs
@@ -1,18 +1,4 @@
-use crate::common::{CommonOpt, Named};
-use structopt::StructOpt;
-
-#[derive(Debug, StructOpt)]
-pub enum RecycleOpt {
-    #[structopt(name = "list")]
-    /// List objects that are in the recycle bin
-    List(CommonOpt),
-    #[structopt(name = "get")]
-    /// Display an object from the recycle bin
-    Get(Named),
-    #[structopt(name = "revive")]
-    /// Revive a recycled object into a live (accessible) state - this is the opposite of "delete"
-    Revive(Named),
-}
+use crate::RecycleOpt;
 
 impl RecycleOpt {
     pub fn debug(&self) -> bool {
