+++ kanidm_unix_int/src/cache_invalidate.rs
@@ -20,15 +20,11 @@ use kanidm_unix_common::client::call_daemon;
 use kanidm_unix_common::unix_config::KanidmUnixdConfig;
 use kanidm_unix_common::unix_proto::{ClientRequest, ClientResponse};
 
-#[derive(Debug, StructOpt)]
-struct ClientOpt {
-    #[structopt(short = "d", long = "debug")]
-    debug: bool,
-}
+include!("./opt/cache_invalidate.rs");
 
 #[tokio::main]
 async fn main() {
-    let opt = ClientOpt::from_args();
+    let opt = CacheInvalidateOpt::from_args();
     if opt.debug {
         ::std::env::set_var("RUST_LOG", "kanidm=debug,kanidm_client=debug");
     } else {
