+++ kanidm_unix_int/src/opt/cache_clear.rs
@@ -1,0 +1,9 @@
+#[derive(Debug, StructOpt)]
+struct CacheClearOpt {
+    #[structopt(short = "d", long = "debug")]
+    debug: bool,
+    #[structopt(long = "really")]
+    really: bool,
+}
+
+
