+++ kanidm_unix_int/src/opt/cache_invalidate.rs
@@ -1,0 +1,7 @@
+
+#[derive(Debug, StructOpt)]
+struct CacheInvalidateOpt {
+    #[structopt(short = "d", long = "debug")]
+    debug: bool,
+}
+
