+++ kanidm_unix_int/src/opt/ssh_authorizedkeys.rs
@@ -1,0 +1,8 @@
+
+#[derive(Debug, StructOpt)]
+struct SshAuthorizedOpt {
+    #[structopt(short = "d", long = "debug")]
+    debug: bool,
+    #[structopt()]
+    account_id: String,
+}
