+++ kanidm_unix_int/src/opt/unixd_status.rs
@@ -1,0 +1,6 @@
+#[derive(Debug, StructOpt)]
+struct UnixdStatusOpt {
+    #[structopt(short = "d", long = "debug")]
+    debug: bool,
+}
+
